backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
qrcode[pil]==7.4.2
pillow==10.2.0
segno==1.6.0
slowapi==0.1.9
python-multipart==0.0.6
